decoupled page component into a wrapper for Homepage, Chapter, & Chronology
